move config to its own package
